Seperate menu from main, edit settings -Also keys showing modified
