+++ unogame/utils/settings.py
@@ -4,10 +4,10 @@ from pygame.locals import *
 
 class Setting:
     
-    def __init__(self, keys, font, screen, visible):
+    def __init__(self, keys, font, screen):
         self.items = [
                       ["Window Size", "Key Configuration", "Color Blindness Mode", "Reset Settings"],
-                      ["size 1", "size 2", "Fullscreen"],
+                      ["800 x 600", "size 2", "Fullscreen"],
                       keys,
                       ["Deuteranopia(Red-Green)", "Tritanopia(Blue-Yellow)", "None"],
                       ]
@@ -19,20 +19,22 @@ class Setting:
         self.keys = keys
         self.key_font = pygame.font.SysFont(None, 20)
         self.option = 0 
-        self.visible = visible
+        self.visible = [False, 255]
         
         # Clear the screen
         self.screen.fill((0, 0, 0))
 
 
     def draw(self):
-        if self.visible: self.screen.fill((0, 0, 0))
+        if self.visible[0]: self.screen.fill((0, 0, 0))
         self.title_text = self.title_font.render(self.titles[self.option], True, (255, 255, 255))
+        screenW = self.screen.get_width()
+        screenH = self.screen.get_height()
 
         # Draw the title
         self.screen.blit(
             self.title_text,
-            (self.screen.get_width() // 10, self.screen.get_height() // 12),
+            (screenW // 10, screenH // 12),
         )
 
         # Draw the menu items
@@ -52,21 +54,43 @@ class Setting:
                 
             self.screen.blit(
                 text,
-                (self.screen.get_width() // 5, self.screen.get_height() // 4 + i * gap)
+                (screenW // 5, screenH // 4 + i * gap)
             )
         
+        #Draw "Save"
         text = self.font.render("Save", True,
                                 (255, 255, 255) if len(self.items[self.option]) != self.selected else (255, 255, 0)
                                 )
         self.screen.blit(
             text,
-            (self.screen.get_width() // 5, self.screen.get_height() * 10 // 12)
+            (screenW // 5, screenH * 10 // 12)
             )
+        
+        #Show Keys Event
+        if self.visible[0]: 
+            for i, (name, key) in enumerate(self.keys):
+                text = self.key_font.render(
+                        f"{name}: {key}",
+                        True,
+                        (255, 255, 255)
+                    )
+                text.set_alpha(self.visible[1])
+                self.screen.blit(text, (
+                    screenW // 10 * 8,
+                    screenH // 12 + i * 30
+                    ))
+            
+            if self.visible[1] < 10:
+                self.visible[0] = False
+                self.visible[1] = 0
+            else: self.visible[1] -= 5
 
 
 
     def run(self):
         clock = pygame.time.Clock()
+        screenW = self.screen.get_width()
+        screenH = self.screen.get_height()
         while True:
             # Handle events
             for event in pygame.event.get():
@@ -92,27 +116,13 @@ class Setting:
                             self.option = 0
                         elif self.option == 1:
                             self.screenSize(self.selected+1)
-                        
-
+                            screenW = self.screen.get_width()
+                            screenH = self.screen.get_height()
                     elif event.key == K_ESCAPE:
                         pygame.quit()
                         sys.exit()
                     else:
-                        self.visible = not self.visible
-                        if self.visible:
-                            for i, (name, key) in enumerate(self.keys):
-                                text = self.key_font.render(
-                                    f"{name}: {key}",
-                                    True,
-                                    (255, 255, 255),
-                                )
-                                self.screen.blit(
-                                    text,
-                                    (self.screen.get_width() // 10 * 8,
-                                     self.screen.get_height() // 12 + i * 30)
-                                    )
-                        self.visible = not self.visible
-
+                        self.visible = [True, 255]
 
                 elif event.type == MOUSEMOTION:
                     pos = pygame.mouse.get_pos()
@@ -122,20 +132,20 @@ class Setting:
                             text = self.font.render(
                                 f"{item[0]}: {item[1]}",
                                 True,
-                                (255, 255, 255) if i != self.selected else (255, 0, 0)
+                                (255, 255, 255)
                             )
                             gap = 50
                         else:
                             text = self.font.render(
                             item,
                             True,
-                            (255, 255, 255) if i != self.selected else (255, 0, 0),
+                            (255, 255, 255)
                             )
                         
                         rect = text.get_rect()
                         rect.topleft = (
-                            self.screen.get_width() // 5,
-                            self.screen.get_height() // 4 + i * gap
+                            screenW // 5,
+                            screenH // 4 + i * gap
                         )
                         if rect.collidepoint(pos):
                             self.selected = i
@@ -143,8 +153,8 @@ class Setting:
                     save = self.font.render("Save", True, (255, 255, 255))
                     rectSave = save.get_rect()
                     rectSave.topleft = (
-                        self.screen.get_width() // 5,
-                        self.screen.get_height() * 10 // 12
+                        screenW // 5,
+                        screenH * 10 // 12
                     )
                     if rectSave.collidepoint(pos):
                         self.selected = len(self.items[self.option])
@@ -160,19 +170,16 @@ class Setting:
                                 (255, 255, 255)
                             )
                             gap = 50
-                        elif type(item) == type(str):
-                            print(type)
-                            #text = self.font.render(item,True,(255, 255, 255)) 
 
                         rect = text.get_rect()
                         rectSave = save.get_rect()
                         rect.topleft = (
-                            self.screen.get_width() // 5,
-                            self.screen.get_height() // 4 + i * gap
+                            screenW // 5,
+                            screenH // 4 + i * gap
                         )
                         rectSave.topleft = (
-                            self.screen.get_width() // 5,
-                            self.screen.get_height() * 10 // 12
+                            screenW // 5,
+                            screenH * 10 // 12
                         )
 
                         if rect.collidepoint(pos):
@@ -183,6 +190,8 @@ class Setting:
                                 self.screen.fill((0, 0, 0))
                             elif self.option == 1:
                                 self.screenSize(self.selected+1)
+                                screenW = self.screen.get_width()
+                                screenH = self.screen.get_height()
 
                         if rectSave.collidepoint(pos):
                             if self.option == 0 and self.selected == 4:
